requirements.txt
pandas
numpy
folium
requests
python-dotenv